Multi_Container_LangChain/text_to_sql/requirements.txt
fastapi == 0.115.0
uvicorn == 0.30.6
langchain == 0.3.1
langchain_groq
httpx == 0.27.2
python-dotenv ==1.0.1
pytest == 8.3.2